chore: always regenerate ruler files on postinstall
Drop the no-op-when-present guard so postinstall re-applies Ruler on every
install, keeping generated files in sync with .ruler/ source. Rename the
script to ruler-apply.mjs to match (it no longer only runs when missing).
