+++ package.json
@@ -236,7 +236,7 @@
     "lint.rust": "make lint",
     "lint.syncpack": "syncpack list-mismatches",
     "preinstall": "npx only-allow pnpm",
-    "postinstall": "node ./scripts/ruler-apply-if-missing.mjs",
+    "postinstall": "node ./scripts/ruler-apply.mjs",
     "prepare": "simple-git-hooks",
     "pack.core": "node --require ./scripts/runBefore.ts scripts/pack-local-qwik-core.ts",
     "prettier.fix": "prettier --cache --write .",
